[react-modal_v3.x.x] Unify react-modal_v3.*.x definitions and support v3.13.1
