+++ definitions/npm/react-modal_v3.x.x/flow_v0.104.x-/react-modal_v3.x.x.js
@@ -1,4 +1,13 @@
 declare module 'react-modal' {
+  declare type ModalElementProps = {|
+    ref: (node: ?HTMLElement) => mixed,
+    id: string,
+    className: string,
+    style: {| [key: string]: string | number |},
+    onClick: (event: SyntheticMouseEvent<any>) => mixed,
+    onMouseDown: (event: SyntheticMouseEvent<any>) => mixed,
+  |};
+
   declare type ModalProps = {
     isOpen?: boolean,
     portalClassName?: string,
@@ -15,6 +24,7 @@ declare module 'react-modal' {
       overlay?: { [key: string]: string | number, ... },
       ...
     },
+    id?: string,
     className?: string | {
       base: string,
       afterOpen: string,
@@ -27,18 +37,30 @@ declare module 'react-modal' {
       beforeClose: string,
       ...
     },
-    onAfterOpen?: () => void | Promise<void>,
+    onAfterOpen?: ({ overlayEl: HTMLElement, contentEl: HTMLElement, ... }) => void | Promise<void>,
+    onAfterClose?: () => void | Promise<void>,
     onRequestClose?: (SyntheticEvent<>) => void,
     aria?: { [key: string]: string, ... },
     role?: string,
+    data?: { [key: string]: string, ... },
+    testId?: string,
     contentLabel?: string,
     overlayRef?: (node: ?HTMLElement) => mixed,
     containerRef?: (node: ?HTMLElement) => mixed,
+    overlayElement?: (
+      props: ModalElementProps,
+      contentElement: HTMLElement
+    ) => HTMLElement,
+    contentElement?: (
+      props: ModalElementProps,
+      contentElement: HTMLElement
+    ) => HTMLElement,
+    preventScroll?: boolean,
     ...
   };
 
   declare class Modal extends React$Component<ModalProps> {
-    static setAppElement(element: HTMLElement | string | null): void;
+    static setAppElement(element: HTMLCollection<HTMLElement> | NodeList<HTMLElement> | HTMLElement[] | HTMLElement | string | null): void;
   }
 
   declare module.exports: typeof Modal;
